Survey form ( for practice )
In this form I learn multiple things like I learn about 
fieldset element for the first time, which we can use for creating section in form, 
legend element which is used to give title to an fieldset element. 
And also learn how to connect input and label element using for tag. 
Now I know the importance of name tag and value tag in radio button.